Crate scratchpad

This crate provides in-memory representation of Aptos core data structures used by the executor.

Structs

FrozenSparseMerkleTree

In the entire lifetime of this, in-mem nodes won’t be dropped because a reference to the oldest SMT is held inside.

SparseMerkleTree

The Sparse Merkle Tree implementation.

Enums

StateStoreStatus

AccountStatus describes the result of querying an account from this SparseMerkleTree.

Traits

ProofRead

A type that implements ProofRead can provide proof for keys in persistent storage.