Trait scratchpad::ProofRead

pub trait ProofRead: Sync {
    fn get_proof(&self, key: HashValue) -> Option<&SparseMerkleProofExt>;
}

A type that implements ProofRead can provide proof for keys in persistent storage.

Required Methods

fn get_proof(&self, key: HashValue) -> Option<&SparseMerkleProofExt>

Gets verified proof for this key in persistent storage.