Crate scratchpad

This crate provides in-memory representation of Diem core data structures used by the executor.

Modules

test_utils

Structs

SparseMerkleTree
The Sparse Merkle Tree implementation.

Enums

AccountStatus
AccountStatus describes the result of querying an account from this SparseMerkleTree.

Traits

ProofRead
A type that implements ProofRead can provide proof for keys in persistent storage.